Physical round expression supports two parameters

+++ datafusion/expr/src/function.rs
@@ -612,8 +612,8 @@ pub fn signature(fun: &BuiltinScalarFunction) -> Signature {
         ),
         BuiltinScalarFunction::Round => Signature::one_of(
             vec![
-                TypeSignature::Exact(vec![DataType::Float64, DataType::Int64]),
-                TypeSignature::Exact(vec![DataType::Float32, DataType::Int64]),
+                TypeSignature::Exact(vec![DataType::Float64, DataType::Int32]),
+                TypeSignature::Exact(vec![DataType::Float32, DataType::Int32]),
                 TypeSignature::Exact(vec![DataType::Float64]),
                 TypeSignature::Exact(vec![DataType::Float32]),
             ],
